fstab: Allow recovery to use the misc partition
Android Nougat changed how it communicates with the recovery
partition. Until now the recovery commands had been written
to the /cache partition, but this has changed. Now recovery
writes to the /misc partition.

Fortunately for kitakami devices and beyond we actually have
an empty and unused /misc partition we can use.

NOTE: avoid

      E:Cannot load volume /misc!
      checked on Marshmallow recovery

Signed-off-by: Adam Farden <adam@farden.cz>
Signed-off-by: David Viteri <davidteri91@gmail.com>